security-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
boto3==1.34.144
redis==5.0.7
celery[redis]==5.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
pip-audit==2.7.3

# dev / test
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
fakeredis==2.23.2
